对于CK进行了一个简易支持

+++ src/main/java/org/guohai/javasqlweb/beans/TablesNameBean.java
@@ -19,5 +19,5 @@ public class TablesNameBean {
     /**
      * 表大小
      */
-    private Integer tableRows;
+    private Long tableRows;
 }
